backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
joblib==1.0.1
requests==2.26.0
python-dotenv==0.19.0
geopandas==0.9.0
matplotlib==3.4.3
seaborn==0.11.2
gunicorn==21.2.0
psycopg2-binary==2.9.1
redis==4.1.0
